chore: move WASM array-call code to arch module
This change moves the details of calling the array-call trampoline to the `arch` module, since jumping to user mode is very architecture specific.
